MODRS-135: Support holdings-storage 6.0, item-storage 10.0
mod-inventory-storage has new major interface versions:

instance-storage 9.0
holdings-storage 6.0
item-storage 10.0

mod-remote-storage only uses the holdings-storage and item-storage interfaces, it doesn't use the instance-storage interface.

mod-remote-storage is not affected because the incompatible change is in the DELETE all APIs only that is not used by mod-remote-storage.

Only the okapiInterfaces need to be bumped in "requires" section of ModuleDescriptor-template.json